added default terrain for terrain.py, added hill functions in map.py, added space around hills, cavemap carver now adds land instead of water

+++ ants/mapgen/util.py
@@ -22,7 +22,7 @@ class Point(collections.namedtuple('Point', ['x', 'y'])):
         return Point(self.x-other.x,self.y-other.y)
     
     def __mul__(self,number):
-        return Point(self.x*number,self.y*number)
+        return Point(int(self.x*number),int(self.y*number))
     
     def random_upto(self):
         """Returns a random point less than self"""
